FFWP#1

+++ forage-midas/src/main/java/com/jpmorgan/midascore/domain/Balance.java
@@ -1,32 +1,36 @@
 package com.jpmorgan.midascore.domain;
 
+import java.math.BigDecimal;
+
 /**
  * Balance — DTO for the GET /balance REST endpoint response.
  *
  * When BalanceController handles GET /balance?userId=..., it returns an
  * instance of this class serialized as JSON:
  *
- *   { "userId": "waldorf", "amount": 950.00 }
+ *   { "userId": "waldorf", "amount": 950.0000 }
+ *
+ * UP-1 Change: amount migrated from double → BigDecimal (PH-03)
  *
  * CRITICAL: DO NOT modify the toString() method.
  *           The test suite (TaskFiveTests) uses toString() for verification.
- *           The exact format must be: Balance[userId=waldorf, amount=950.0]
+ *           The exact format must be: Balance[userId=waldorf, amount=950.0000]
  *
  * Unknown user case:
- *   If userId does not exist in the database, return Balance with amount = 0.0
+ *   If userId does not exist in the database, return Balance with amount = 0.0000
  *   (200 OK response — NOT a 404).
  */
 public class Balance {
 
     private String userId;
-    private double amount;
+    private BigDecimal amount;
 
     // ─── Constructors ───────────────────────────────────────────────────────
 
     /** Required for JSON serialization */
     public Balance() {}
 
-    public Balance(String userId, double amount) {
+    public Balance(String userId, BigDecimal amount) {
         this.userId = userId;
         this.amount = amount;
     }
@@ -41,11 +45,11 @@ public class Balance {
         this.userId = userId;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
